src/requirements.txt
asyncpg==0.20.0
databases[postgresql]==0.2.6
fastapi==0.46.0
PyJWT==1.7.1
SQLAlchemy==1.3.12
uvicorn==0.11.1

# dev
pytest==5.3.2
requests==2.22.0
